Convert an ODK template to the OCA data sharing dictionary standard

Convert an ODK template to the OCA data sharing dictionary standard

Usage

dict_from_odk(survey, choices, col_labels = NULL)

Arguments

survey

A data.frame reflecting the 'survey' sheet within an ODK template

choices

A data.frame reflecting the 'choices' sheet within an ODK template

col_labels

(Optional) Name of the column found in both sheets 'survey' and 'choices' giving variable labels, e.g. "label::English". Defaults to NULL in which case the first column starting with prefix "label" is used for the label column.

Value

A tibble-style data frame representing a data dictionary formatted to the OCA data sharing standard

Examples

path_data <- system.file("extdata", package = "datadict")
path_odk_template <- file.path(path_data, "WHOVA2016_v1_5_3_ODK.xlsx")

odk_survey <- readxl::read_xlsx(path_odk_template, sheet = "survey")
odk_choices <- readxl::read_xlsx(path_odk_template, sheet = "choices")

dict_from_odk(odk_survey, odk_choices)
#> # A tibble: 593 × 17
#>    variabl…¹ short…² type  choices origin status indir…³ hint_…⁴ relev…⁵ requi…⁶
#>    <chr>     <chr>   <chr> <chr>   <chr>  <chr>  <chr>   <chr>   <chr>   <chr>  
#>  1 Id10002   Is thi… Code… high, … origi… shared NA      Should… NA      yes    
#>  2 Id10003   Is thi… Code… high, … origi… shared NA      Should… NA      yes    
#>  3 Id10004   During… Code… wet, W… origi… shared NA      NA      NA      yes    
#>  4 Id10007   What i… Free… NA      origi… shared NA      NA      NA      NA     
#>  5 Id10007a  What i… Code… female… origi… shared NA      NA      NA      yes    
#>  6 Id10008   What i… Code… parent… origi… shared NA      First … NA      yes    
#>  7 Id10009   Did yo… Code… yes, Y… origi… shared NA      NA      NA      yes    
#>  8 Id10010   Name o… Free… NA      origi… shared NA      NA      NA      yes    
#>  9 Id10010a  Age of… Nume… NA      origi… shared NA      NA      NA      yes    
#> 10 Id10010b  Sex of… Code… female… origi… shared NA      NA      NA      yes    
#> # … with 583 more rows, 7 more variables: notes <lgl>, appearance <chr>,
#> #   calculation <chr>, default <lgl>, constraint <chr>,
#> #   constraint_message_english <chr>, read_only <lgl>, and abbreviated variable
#> #   names ¹​variable_name, ²​short_label, ³​indirect_identifier, ⁴​hint_english,
#> #   ⁵​relevant, ⁶​required