Generate a data dictionary

generate_dict() builds a data dictionary from a data frame or tibble. It produces a structured overview of variable metadata, including variable names, labels, data types, value distributions, and any available values or value labels.

generate_dict(
  data,
  columns = NULL,
  n_max = 10L,
  label_fallback = "NO QUESTION LABEL",
  label_width = 100L
)

Arguments

data

A data frame or tibble.

columns

An optional character vector of column names to include in the dictionary. If NULL (default), information for all non-empty columns/ variables are included in the dictionary.

n_max

An integer that controls how detailed numeric reporting should be. If a numeric variable has more unique values than this threshold, it is summarized as a range (e.g., 1 – 100). If it has fewer, each unique value receives its own row in the dictionary. Defaults to 10.

label_fallback

The placeholder text to use when a variable does not have a defined label attribute. Defaults to "NO QUESTION LABEL".

label_width

An integer to truncate long variable labels for better readability. Defaults to 100 characters.

Value

A tibble with the following structural metadata:

variable_name: The column/variable name as it appears in the data source.

variable_label: The descriptive variable label or the default placeholder.

variable_type: The type of data contained within the column/variable (e.g., numeric, character, labelled, factor). Data type is derived using R's class() function.

variable_values: Individual data values or range of data values.

value_labels: The descriptive labels assigned to specific data values.

n_size: The frequency count for each value/label pair.

is_range: A boolean flag indicating if the variable_values column represents a summarized range rather than a discrete value.

Author

Ama Nyame-Mensah

Examples

generate_dict(data = nlsy)
#> # A tibble: 34 × 7
#>    variable_name variable_label    variable_type variable_values    value_labels
#>    <chr>         <chr>             <chr>         <chr>              <chr>       
#>  1 CID           NO QUESTION LABEL numeric       303 - 1266703      NA          
#>  2 CID           NO QUESTION LABEL numeric       NA                 system miss…
#>  3 race          NO QUESTION LABEL character     Black              Black       
#>  4 race          NO QUESTION LABEL character     Hispanic           Hispanic    
#>  5 race          NO QUESTION LABEL character     Non-Black,Non-His… Non-Black,N…
#>  6 race          NO QUESTION LABEL character     NA                 system miss…
#>  7 gender        NO QUESTION LABEL numeric       0                  0           
#>  8 gender        NO QUESTION LABEL numeric       1                  1           
#>  9 gender        NO QUESTION LABEL numeric       NA                 system miss…
#> 10 birthord      NO QUESTION LABEL numeric       1                  1           
#> # ℹ 24 more rows
#> # ℹ 2 more variables: n_size <int>, is_range <lgl>

# example shortening variable labels to 25 characters
generate_dict(data = gss, label_width = 25)
#> # A tibble: 772 × 7
#>    variable_name variable_label       variable_type variable_values value_labels
#>    <chr>         <chr>                <chr>         <chr>           <chr>       
#>  1 id            Respondent ID number numeric       1 - 4150        NA          
#>  2 id            Respondent ID number numeric       NA              system miss…
#>  3 year          GSS year for this r… labelled      -100            iap         
#>  4 year          GSS year for this r… labelled      -99             no answer   
#>  5 year          GSS year for this r… labelled      -98             don't know  
#>  6 year          GSS year for this r… labelled      -97             skipped on …
#>  7 year          GSS year for this r… labelled      -96             see codebook
#>  8 year          GSS year for this r… labelled      -95             uncodeable  
#>  9 year          GSS year for this r… labelled      -94             not imputab…
#> 10 year          GSS year for this r… labelled      -90             refused     
#> # ℹ 762 more rows
#> # ℹ 2 more variables: n_size <int>, is_range <lgl>

# example selecting a subset of variables from a tibble
generate_dict(
  data = gss,
  columns = c("year", "age", "sexnow1", "hispanic", "race", "marital")
)
#> # A tibble: 168 × 7
#>    variable_name variable_label       variable_type variable_values value_labels
#>    <chr>         <chr>                <chr>         <chr>           <chr>       
#>  1 year          GSS year for this r… labelled      -100            iap         
#>  2 year          GSS year for this r… labelled      -99             no answer   
#>  3 year          GSS year for this r… labelled      -98             don't know  
#>  4 year          GSS year for this r… labelled      -97             skipped on …
#>  5 year          GSS year for this r… labelled      -96             see codebook
#>  6 year          GSS year for this r… labelled      -95             uncodeable  
#>  7 year          GSS year for this r… labelled      -94             not imputab…
#>  8 year          GSS year for this r… labelled      -90             refused     
#>  9 year          GSS year for this r… labelled      -80             not availab…
#> 10 year          GSS year for this r… labelled      -70             not availab…
#> # ℹ 158 more rows
#> # ℹ 2 more variables: n_size <int>, is_range <lgl>