Nursing Workforce Survey Data

These data contain a random sample of 5,000 records from the 2008 National Sample Survey of Registered Nurses (SPSS Format) and the National Sample Survey of Nurse Practitioners, specifically formatted for use in this package. For more information about the Nursing Workforce Survey Dataset, please visit: https://data.hrsa.gov/topics/health-workforce/nchwa/nursing-workforce-survey-data.

nursing_work

Format

A tibble with 5,000 rows and 865 columns:

Q1

Actively licensed RN

Q2STATE

State of first RN license

Q2YEAR

Year of first RN license

Q4MONTH

Month graduated initial RN ed program

Q4YEAR

Year graduated initial RN ed program

Q5STATE

State initial RN ed program located

Q5COU

Country initial RN ed program located

Q6_0

No degree prior to RN ed program

Q6_1

AD prior to RN ed program

Q6_2

Bachelor's prior to RN ed program

Q6_3

Master's prior to RN ed program

Q6_4

Doctorate prior to RN ed program

Q6_5

Other degree prior to RN ed program

Q7

Field for highest degree before RN ed

Q8

Licensed LPN/LVN

Q9_0

No health-related position before RN ed

Q9_1

Nursing Aide/Asst before RN ed

Q9_2

Home health aide/asst before RN ed

Q9_3

LPN/LVN before RN ed

Q9_4

EMT before RN ed

Q9_5

Medical asst before RN ed

Q9_6

Dental asst before RN ed

Q9_7

Allied Health/Lab tech before RN ed

Q9_8

Health care manager before RN ed

Q9_9

Health care clerk before RN ed

Q9_10

Military med corps before RN ed

Q9_11

MD before RN ed

Q9_12

Midwife before RN ed

Q9_13

Other health-related position before RN ed

Q10_1

Financed initial RN ed - Health-related employment

Q10_2

Financed initial RN ed - Nonhealth-related employment

Q10_3

Financed initial RN ed - Other household members

Q10_4

Financed initial RN ed - Personal savings

Q10_5

Financed initial RN ed - Other family resources

Q10_6

Financed initial RN ed - Employer tuition reimbursement

Q10_7

Financed initial RN ed - Federal scholarship/grant

Q10_8

Financed initial RN ed - Federal assisted loan

Q10_9

Financed initial RN ed - Other loan

Q10_10

Financed initial RN ed - State/local gov't scholarship/grant

Q10_11

Financed initial RN ed - Non-gov't scholarship/grant

Q10_12

Financed initial RN ed - Other resources

Q11A_0

No emergency preparedness training received/provided

Q11A_1

Training received/provided - Chemical accident/attack

Q11A_2

Training received/provided - Nuclear/radiological accident/attack

Q11A_3

Training received/provided - Infectious disease epidemics

Q11A_4

Training received/provided - Biological accident/attack

Q11A_5

Training received/provided - Natural disaster

Q11A_6

Training received/provided - Other public health emergencies

Q11BREC

Total number of training hours received

Q11BPROV

Total number of training hours provided

Q11C

Best prepared emergency area

Q12

Know disaster/emergency plan at work

Q13

Additional degrees after RN program

Q14

Completed APRN program

Q14A_A

Prepared as NP

Q14B_A

NP program length

Q14C_A

Highest NP program credential

Q14D_A

Year received NP credential

Q14E_A

NP certification national org

Q14E2_A

NP certification required for job

Q14F_A

NP State Board license/recognition

Q14F2_A

NP State Board license/recognition required for job

Q14A_B

Prepared as CNS

Q14B_B

CNS program length

Q14C_B

Highest CNS program credential

Q14D_B

Year received CNS credential

Q14E_B

CNS certification national org

Q14E2_B

CNS certification required for job

Q14F_B

CNS State Board license/recognition

Q14F2_B

CNS State Board license/recognition required for job

Q14A_C

Prepared as NM

Q14B_C

NM program length

Q14C_C

Highest NM program credential

Q14D_C

Year received NM credential

Q14E_C

NM certification national org

Q14E2_C

NM certification required for job

Q14F_C

NM State Board license/recognition

Q14F2_C

NM Sate Board license/recognition required for job

Q14A_D

Prepared as NA

Q14B_D

NA program length

Q14C_D

Highest NA program credential

Q14D_D

Year received NA credential

Q14E_D

NA certification national org

Q14E2_D

NA certification required for job

Q14F_D

NA State Board license/recognition

Q14F2_D

NA State Board license/recognition required for job

Q15

Currently enrolled in formal ed program 2008

Q16

Type of formal ed program 2008

Q17A

FT/PT student 2008

Q17B

Percent distance-based courses 2008

Q18

Type of degree/certificate 2008

Q19

Currently employed in nursing 2008

Q20

Active RN license required

Q21STATE

State principal RN position

Q22

Principal RN position type 2008

Q23

Principal RN position setting 2008

Q24

Principal RN job title 2008

Q25

FT/PT principal RN position in 2008

Q26

Months worked principal RN position 2008

Q27A

Number hours worked

Q27B

Number on-call hours

Q27C

Number overtime hours

Q27D

Number mandatory overtime hours

Q27E

Number paid/unpaid stand-by hours

Q27F

Number hours paid stand-by rate

Q28A

Percent patient care & charting

Q28B

Percent non-nursing tasks

Q28C

Percent consultation

Q28D

Percent supervision & management

Q28E

Percent administration

Q28F

Percent research

Q28G

Percent teaching, orientating students

Q28H

Percent other activities

Q29A_1

Level of care - General/specialty inpatient

Q29A_2

Level of care - Critical/intensive care

Q29A_3

Level of care - Step down, transitional care

Q29A_4

Level of care - Sub-acute care

Q29A_5

Level of care - Emergency care

Q29A_6

Level of care - Urgent care

Q29A_7

Level of care -Rehabilitation

Q29A_8

Level of care - Long-term care

Q29A_9

Level of care - Surgery

Q29A_10

Level of care - Ambulatory care

Q29A_11

Level of care - Ancillary care

Q29A_12

Level of care - Home health

Q29A_13

Level of care - Public health

Q29A_14

Level of care - Education

Q29A_15

Level of care - Business

Q29A_16

Level of care - Research

Q29A_17

Level of care - Other specify

Q29B

Primary patient population

Q29C_1

No patient care principal position

Q29C_2

Clinical specialty - General surgery

Q29C_3

Clinical specialty - Critical care

Q29C_4

Clinical specialty - Cardiac care

Q29C_5

Clinical specialty - Chronic care

Q29C_6

Clinical specialty - Dermatology

Q29C_7

Clinical specialty - Emergency/trauma

Q29C_8

Clinical specialty - Gastrointestinal

Q29C_9

Clinical specialty - Gynecology

Q29C_10

Clinical specialty - Hospice

Q29C_11

Clinical specialty - Infectious disease

Q29C_12

Clinical specialty - Labor & delivery

Q29C_13

Clinical specialty - Neurological

Q29C_14

Clinical specialty - Obstetrics

Q29C_15

Clinical specialty - Occupational health

Q29C_16

Clinical specialty - Oncology

Q29C_17

Clinical specialty - Primary care

Q29C_18

Clinical specialty - Psychiatric/mental health

Q29C_19

Clinical specialty - Pulmonary/respiratory

Q29C_20

Clinical specialty - Radiology

Q29C_21

Clinical specialty - Renal/dialysis

Q29C_22

Clinical specialty - No specific area

Q29C_23

Clinical specialty - Other specialty

Q30

Pre-tax earnings principal RN position 2008

Q31

Represented by labor union 2008

Q32A

Intentions regarding principal RN position

Q32B

Plans to remain in nursing profession

Q33

Hold other paid RN positions 2008

Q34_1

Secondary RN position - Employee at org/facility

Q34_2

Secondary RN position - Employed thru employment agency

Q34_3

Secondary RN position - Employed as traveling nurse

Q34_4

Secondary RN position - Self-employed, working as-needed

Q35_1

Secondary RN position - Hospital setting

Q35_2

Secondary RN position - Nursing home/Extended care setting

Q35_3

Secondary RN position - Academic ed program setting

Q35_4

Secondary RN position - Home health setting

Q35_5

Secondary RN position - Public/community health setting

Q35_6

Secondary RN position - School health service setting

Q35_7

Secondary RN position - Occupational health setting

Q35_8

Secondary RN position - Ambulatory care setting

Q35_9

Secondary RN position - Insurance claims/benefits setting

Q35_10

Secondary RN position - Telehealth, telenursing or call center setting

Q35_11

Secondary RN position - Other specify (setting)

Q36A_1

Weeks per year (additional RN job #'1)

Q36B_1

Avg hours per week (additional RN job #'1)

Q36C_1

State (additional RN job #'1)

Q36A_2

Weeks per year (additional RN job #' 2)

Q36B_2

Avg hours per week (additional RN job #'2)

Q36C_2

State (additional RN job #'2)

Q36A_3

Weeks per year (all other RN jobs)

Q36B_3

Avg hours per week (all other RN jobs)

Q37

Pre-tax earnings other RN positions 2008

Q38

Paid RN work intentions

Q39A

Months actively looking for paid RN work

Q40

Looking for FT/PT position

Q41

Plan to return to paid RN work

Q42

Last employed/self-employed as RN

Q43_1

Reason not RN - Retired

Q43_2

Reason not RN - Personal

Q43_3

Reason not RN - Burnout

Q43_4

Reason not RN - Work stress

Q43_5

Reason not RN - Scheduling

Q43_6

Reason not RN - Physical demands

Q43_7

Reason not RN - Disability

Q43_8

Reason not RN - Illness

Q43_9

Reason not RN - Inadequate staffing

Q43_10

Reason not RN - Low salary

Q43_11

Reason not RN - Out-of-date skills

Q43_12

Reason not RN - Liability concerns

Q43_13

Reason not RN - Lack of collaboration

Q43_14

Reason not RN - Unable to be RN on professional level

Q43_15

Reason not RN - Lack advancement opportunities

Q43_16

Reason not RN - Lack good leadership

Q43_17

Reason not RN - Career change

Q43_18

Reason not RN - Difficulty finding RN position

Q43_19

Reason not RN - Travel

Q43_20

Reason not RN - Volunteering in nursing

Q43_21

Reason not RN - Back to school

Q43_22

Reason not RN - Other reason

Q44

Non-RN employment 2008

Q45

Non-RN employment health-related 2008

Q46A

Field of non-RN principal position

Q46B

Non-RN position job title

Q47

Months worked non-RN position

Q48

Avg hours per week non-RN position

Q49

Pre-tax earnings non-RN position

Q50

Years worked as RN (since 1st U.S. RN license)

Q51

Left nursing 1 or more years since becoming RN

Q51YRS

Total years not working in nursing

Q52

Employed as RN 1 year ago

Q53

FT/PT principal RN position in 2007

Q54

Employed with current employer in 2007

Q55STATE

State 2007 principal RN position

Q56_1

Job change - Burnout

Q56_2

Job change - Work stress

Q56_3

Job change - Interest in other position

Q56_4

Job change - Lack advancement

Q56_5

Job change - Lack collaboration

Q56_6

Job change - Lack leadership

Q56_7

Job change - Career advancement

Q56_8

Job change - Inadequate staffing

Q56_9

Job change - Differences with colleagues

Q56_10

Job change - Physical demands

Q56_11

Job change - Do kind of nursing I like

Q56_12

Job change - Better pay/benefits

Q56_13

Job change - Scheduling

Q56_14

Job change - Relocation

Q56_15

Job change - Shift in positions/reorganization

Q56_16

Job change - Lay offs/downsizing

Q56_17

Job change - Sign-on bonus offered

Q56_18

Job change - Personal

Q56_19

Job change - Back to school

Q56_20

Job change - Retired

Q56_21

Job change - Disability

Q56_22

Job change - Illness

Q56_23

Job change - Other

Q57

Employment setting 2007 principal RN position

Q58

Satisfaction with principal/most recent RN job

Q59STATE

Current state residence

Q60

Reside in same city/town a year ago

Q61STATE

State 2007 residence

Q62

Gender

Q63

Year of birth

Q64A

Latino/Hispanic ethnicity

Q65_1

No other languages

Q65_2

Spanish language

Q65_3

Filipino language

Q65_5

French language

Q66

Marital status

Q67_1

No dependents at home

Q67_2

Children less than 6 years at home

Q67_3

Children 6 to 18 years at home

Q67_4

Other adults at home

Q67_5

Others living elsewhere

Q68

Current, pre-tax annual household income

Q70A

Current National RN certifications

Q70B_0

No current skill-based certifications

Q70B_1

Life Support certification

Q70B_2

Resuscitation certification

Q70B_3

Emergency medicine certification

Q70B_4

Trauma certification

Q70B_5

1st Other skill-based certification

Q70B_6

2nd Other skill-based certification

Q70C_31

Nurse Midwife (CNM) certification

Q70C_32

Nurse Anesthetist (CRNA) certification

Q70D_0

Other National RN certifications

SOURCE

Data collection type

SUBMISSION_DATE

Submission date

Q9_14

Student Nurse Externship

Q29C_24

Clinical specialty - Plastic/Cosmetic Surgery

Q29C_25

Clinical specialty - Orthopedic

Q29C_26

Clinical specialty - Endocrinology

Q29C_27

Clinical specialty - Urology

Q9_16

Health Educator before RN ed

Q9_15

Mental Health/Psychiatric work before RN ed

Q29A_18

Level of care - Informatics

Q29A_19

Level of care - Flight nurse/med air transport

Q29C_28

Clinical specialty - Infusion/IV

Q29C_29

Clinical specialty - Ophthalmology

Q29C_30

Clinical specialty - Otolaryngology/ENT

Q35_12

Secondary RN position - Military/Reserves/National Guard setting

Q35_13

Secondary RN position - Medical Legal setting

Q35_14

Secondary RN position - Consultant/Contractor setting

Q56_24

Job change - Contract ended

IMPQ6_0

Imputation flag for Q6_0

IMPQ8

Imputation flag for Q8

IMPQ9_0

Imputation flag for Q9_0

IMPQ11A_0

Imputation flag for Q11A_0

IMPQ12

Imputation flag for Q12

IMPQ15

Imputation flag for Q15

IMPQ44

Imputation flag for Q44

IMPQ60

Imputation flag for Q60

IMPQ62

Imputation flag for Q62

IMPQ67_1

Imputation flag for Q67_1

IMPQ63

Imputation flag for Q63

IMPQ4YEAR

Imputation flag for Q4YEAR

IMPQ2YEAR

Imputation flag for Q2YEAR

IMPQ50

Imputation flag for Q50

IMPQ9_1

Imputation flag for Q9_1

IMPQ9_2

Imputation flag for Q9_2

IMPQ9_3

Imputation flag for Q9_3

IMPQ9_4

Imputation flag for Q9_4

IMPQ9_5

Imputation flag for Q9_5

IMPQ9_6

Imputation flag for Q9_6

IMPQ9_7

Imputation flag for Q9_7

IMPQ9_8

Imputation flag for Q9_8

IMPQ9_9

Imputation flag for Q9_9

IMPQ9_10

Imputation flag for Q9_10

IMPQ9_11

Imputation flag for Q9_11

IMPQ9_12

Imputation flag for Q9_12

IMPQ9_13

Imputation flag for Q9_13

IMPQ9_14

Imputation flag for Q9_14

IMPQ9_15

Imputation flag for Q9_15

IMPQ9_16

Imputation flag for Q9_16

IMPQ10_1

Imputation flag for Q10_1

IMPQ10_2

Imputation flag for Q10_2

IMPQ10_3

Imputation flag for Q10_3

IMPQ10_4

Imputation flag for Q10_4

IMPQ10_5

Imputation flag for Q10_5

IMPQ10_6

Imputation flag for Q10_6

IMPQ10_7

Imputation flag for Q10_7

IMPQ10_8

Imputation flag for Q10_8

IMPQ10_9

Imputation flag for Q10_9

IMPQ10_10

Imputation flag for Q10_10

IMPQ10_11

Imputation flag for Q10_11

IMPQ10_12

Imputation flag for Q10_12

IMPQ68

Imputation flag for Q68

IMPQ52

Imputation flag for Q52

IMPQ58

Imputation flag for Q58

IMPQ64A

Imputation flag for Q64A

IMPQ64B_1

Imputation flag for Q64B_1

IMPQ64B_2

Imputation flag for Q64B_2

IMPQ64B_3

Imputation flag for Q64B_3

IMPQ64B_4

Imputation flag for Q64B_4

IMPQ64B_5

Imputation flag for Q64B_5

IMPQ65_1

Imputation flag for Q65_1

IMPQ65_2

Imputation flag for Q65_2

IMPQ65_3

Imputation flag for Q65_3

IMPQ65_4

Imputation flag for Q65_4

IMPQ65_5

Imputation flag for Q65_5

IMPQ65_6

Imputation flag for Q65_6

IMPQ65_7

Imputation flag for Q65_7

IMPQ65_8

Imputation flag for Q65_8

IMPQ65_9

Imputation flag for Q65_9

IMPQ65_10

Imputation flag for Q65_10

IMPQ65_11

Imputation flag for Q65_11

IMPQ65_12

Imputation flag for Q65_12

IMPQ65_13

Imputation flag for Q65_13

IMPQ65_14

Imputation flag for Q65_14

IMPQ65_15

Imputation flag for Q65_15

IMPQ65_16

Imputation flag for Q65_16

IMPQ65_17

Imputation flag for Q65_17

IMPQ65_18

Imputation flag for Q65_18

IMPQ65_19

Imputation flag for Q65_19

IMPQ65_20

Imputation flag for Q65_20

IMPQ65_21

Imputation flag for Q65_21

IMPQ65_22

Imputation flag for Q65_22

IMPQ65_23

Imputation flag for Q65_23

IMPQ65_24

Imputation flag for Q65_24

IMPQ65_25

Imputation flag for Q65_25

IMPQ65_26

Imputation flag for Q65_26

IMPQ65_27

Imputation flag for Q65_27

IMPQ67_2

Imputation flag for Q67_2

IMPQ67_3

Imputation flag for Q67_3

IMPQ67_4

Imputation flag for Q67_4

IMPQ67_5

Imputation flag for Q67_5

IMPQ3

Imputation flag for Q3

IMPQ4MONTH

Imputation flag for Q4MONTH

IMPQ6_1

Imputation flag for Q6_1

IMPQ6_2

Imputation flag for Q6_2

IMPQ6_3

Imputation flag for Q6_3

IMPQ6_4

Imputation flag for Q6_4

IMPQ6_5

Imputation flag for Q6_5

IMPQ7

Imputation flag for Q7

IMPQ11A_1

Imputation flag for Q11A_1

IMPQ11A_2

Imputation flag for Q11A_2

IMPQ11A_3

Imputation flag for Q11A_3

IMPQ11A_4

Imputation flag for Q11A_4

IMPQ11A_5

Imputation flag for Q11A_5

IMPQ11A_6

Imputation flag for Q11A_6

IMPQ11BPROV

Imputation flag for Q11BPROV

IMPQ11BREC

Imputation flag for Q11BREC

IMPQ11C

Imputation flag for Q11C

IMPQ13A_A

Imputation flag for Q13A_A

IMPQ13A_B

Imputation flag for Q13A_B

IMPQ13A_C

Imputation flag for Q13A_C

IMPQ13A_D

Imputation flag for Q13A_D

IMPQ13A_E

Imputation flag for Q13A_E

IMPQ13A_F

Imputation flag for Q13A_F

IMPQ13A_G

Imputation flag for Q13A_G

IMPQ13A_H

Imputation flag for Q13A_H

IMPQ13A_I

Imputation flag for Q13A_I

IMPQ13A_J

Imputation flag for Q13A_J

IMPQ13B_C

Imputation flag for Q13B_C

IMPQ13B_D

Imputation flag for Q13B_D

IMPQ13B_E

Imputation flag for Q13B_E

IMPQ13B_F

Imputation flag for Q13B_F

IMPQ13B_G

Imputation flag for Q13B_G

IMPQ13B_H

Imputation flag for Q13B_H

IMPQ13B_I

Imputation flag for Q13B_I

IMPQ13B_J

Imputation flag for Q13B_J

IMPQ13C_F

Imputation flag for Q13C_F

IMPQ13C_G

Imputation flag for Q13C_G

IMPQ13C_H

Imputation flag for Q13C_H

IMPQ13C_I

Imputation flag for Q13C_I

IMPQ13C_J

Imputation flag for Q13C_J

IMPQ13D_A

Imputation flag for Q13D_A

IMPQ13D_B

Imputation flag for Q13D_B

IMPQ13D_C

Imputation flag for Q13D_C

IMPQ13D_D

Imputation flag for Q13D_D

IMPQ13D_E

Imputation flag for Q13D_E

IMPQ13D_F

Imputation flag for Q13D_F

IMPQ13D_G

Imputation flag for Q13D_G

IMPQ13D_H

Imputation flag for Q13D_H

IMPQ13D_I

Imputation flag for Q13D_I

IMPQ13D_J

Imputation flag for Q13D_J

IMPQ13F_A

Imputation flag for Q13F_A

IMPQ13F_B

Imputation flag for Q13F_B

IMPQ13F_C

Imputation flag for Q13F_C

IMPQ13F_D

Imputation flag for Q13F_D

IMPQ13F_E

Imputation flag for Q13F_E

IMPQ13F_F

Imputation flag for Q13F_F

IMPQ13F_G

Imputation flag for Q13F_G

IMPQ13F_H

Imputation flag for Q13F_H

IMPQ13F_I

Imputation flag for Q13F_I

IMPQ13F_J

Imputation flag for Q13F_J

IMPQ16

Imputation flag for Q16

IMPQ17A

Imputation flag for Q17A

IMPQ17B

Imputation flag for Q17B

IMPQ18

Imputation flag for Q18

IMPQ20

Imputation flag for Q20

IMPQ22

Imputation flag for Q22

IMPQ25

Imputation flag for Q25

IMPQ26

Imputation flag for Q26

IMPQ27C

Imputation flag for Q27C

IMPQ27B

Imputation flag for Q27B

IMPQ27A

Imputation flag for Q27A

IMPQ27D

Imputation flag for Q27D

IMPQ27E

Imputation flag for Q27E

IMPQ27F

Imputation flag for Q27F

IMPQ28A

Imputation flag for Q28A

IMPQ28B

Imputation flag for Q28B

IMPQ28C

Imputation flag for Q28C

IMPQ28D

Imputation flag for Q28D

IMPQ28E

Imputation flag for Q28E

IMPQ28F

Imputation flag for Q28F

IMPQ28G

Imputation flag for Q28G

IMPQ28H

Imputation flag for Q28H

IMPQ30

Imputation flag for Q30

IMPQ31

Imputation flag for Q31

IMPQ32A

Imputation flag for Q32A

IMPQ32B

Imputation flag for Q32B

IMPQ33

Imputation flag for Q33

IMPQ34_1

Imputation flag for Q34_1

IMPQ34_2

Imputation flag for Q34_2

IMPQ34_3

Imputation flag for Q34_3

IMPQ34_4

Imputation flag for Q34_4

IMPQ35_1

Imputation flag for Q35_1

IMPQ35_2

Imputation flag for Q35_2

IMPQ35_3

Imputation flag for Q35_3

IMPQ35_4

Imputation flag for Q35_4

IMPQ35_5

Imputation flag for Q35_5

IMPQ35_6

Imputation flag for Q35_6

IMPQ35_7

Imputation flag for Q35_7

IMPQ35_8

Imputation flag for Q35_8

IMPQ35_9

Imputation flag for Q35_9

IMPQ35_10

Imputation flag for Q35_10

IMPQ35_11

Imputation flag for Q35_11

IMPQ35_12

Imputation flag for Q35_12

IMPQ35_13

Imputation flag for Q35_13

IMPQ35_14

Imputation flag for Q35_14

IMPQ36A_1

Imputation flag for Q36A_1

IMPQ36A_2

Imputation flag for Q36A_2

IMPQ36A_3

Imputation flag for Q36A_3

IMPQ36B_1

Imputation flag for Q36B_1

IMPQ36B_2

Imputation flag for Q36B_2

IMPQ36B_3

Imputation flag for Q36B_3

IMPQ36C_1CY

Imputation flag for Q36C_1CY

IMPQ36C_2CY

Imputation flag for Q36C_2CY

IMPQ37

Imputation flag for Q37

IMPQ38

Imputation flag for Q38

IMPQ39A

Imputation flag for Q39A

IMPQ40

Imputation flag for Q40

IMPQ41

Imputation flag for Q41

IMPQ43_1

Imputation flag for Q43_1

IMPQ43_2

Imputation flag for Q43_2

IMPQ43_3

Imputation flag for Q43_3

IMPQ43_4

Imputation flag for Q43_4

IMPQ43_5

Imputation flag for Q43_5

IMPQ43_6

Imputation flag for Q43_6

IMPQ43_7

Imputation flag for Q43_7

IMPQ43_8

Imputation flag for Q43_8

IMPQ43_9

Imputation flag for Q43_9

IMPQ43_10

Imputation flag for Q43_10

IMPQ43_11

Imputation flag for Q43_11

IMPQ43_12

Imputation flag for Q43_12

IMPQ43_13

Imputation flag for Q43_13

IMPQ43_14

Imputation flag for Q43_14

IMPQ43_15

Imputation flag for Q43_15

IMPQ43_16

Imputation flag for Q43_16

IMPQ43_17

Imputation flag for Q43_17

IMPQ43_18

Imputation flag for Q43_18

IMPQ43_19

Imputation flag for Q43_19

IMPQ43_20

Imputation flag for Q43_20

IMPQ43_21

Imputation flag for Q43_21

IMPQ43_22

Imputation flag for Q43_22

IMPQ45

Imputation flag for Q45

IMPQ46A

Imputation flag for Q46A

IMPQ46B

Imputation flag for Q46B

IMPQ47

Imputation flag for Q47

IMPQ48

Imputation flag for Q48

IMPQ49

Imputation flag for Q49

IMPQ51

Imputation flag for Q51

IMPQ51YRS

Imputation flag for Q51YRS

IMPQ42

Imputation flag for Q42

IMPQ53

Imputation flag for Q53

IMPQ54

Imputation flag for Q54

IMPQ56_1

Imputation flag for Q56_1

IMPQ56_2

Imputation flag for Q56_2

IMPQ56_3

Imputation flag for Q56_3

IMPQ56_4

Imputation flag for Q56_4

IMPQ56_5

Imputation flag for Q56_5

IMPQ56_6

Imputation flag for Q56_6

IMPQ56_7

Imputation flag for Q56_7

IMPQ56_8

Imputation flag for Q56_8

IMPQ56_9

Imputation flag for Q56_9

IMPQ56_10

Imputation flag for Q56_10

IMPQ56_11

Imputation flag for Q56_11

IMPQ56_12

Imputation flag for Q56_12

IMPQ56_13

Imputation flag for Q56_13

IMPQ56_14

Imputation flag for Q56_14

IMPQ56_15

Imputation flag for Q56_15

IMPQ56_16

Imputation flag for Q56_16

IMPQ56_17

Imputation flag for Q56_17

IMPQ56_18

Imputation flag for Q56_18

IMPQ56_19

Imputation flag for Q56_19

IMPQ56_20

Imputation flag for Q56_20

IMPQ56_21

Imputation flag for Q56_21

IMPQ56_22

Imputation flag for Q56_22

IMPQ56_23

Imputation flag for Q56_23

IMPQ56_24

Imputation flag for Q56_24

IMPQ66

Imputation flag for Q66

IMPQ2STATE

Imputation flag for Q2STATE

IMPQ5STATE

Imputation flag for Q5STATE

IMPQ55STATE

Imputation flag for Q55STATE

IMPQ57

Imputation flag for Q57

IMPQ61STATE

Imputation flag for Q61STATE

IMPQ36C_1

Imputation flag for Q36C_1

IMPQ36C_2

Imputation flag for Q36C_2

AGE

Age of RN in 2007

AGE_GP

Ages grouped

MAR_ADULT

Marital status & adult dependants

MAR_CHLD

Marital status & children

NLANG

Non-English languages spoken category count

RACE

Single races

RAC_ETHN

Race/ethnicity

RACE_GP

Race/ethnicity grouped

HN_YEAR

Year received highest nursing ed

HN_AGE

Age at highest nursing ed

H_YEAR

Year received highest nursing/nursing-related ed

AGE_GRAD

Age at graduation - initial RN ed

GRAD_YR

Initial RN ed graduation year

YR_IN_HN

Years between initial and highest RN ed

YR_SINCE

Years since graduated from initial RN ed program

YRSNC_GP

Years since graduated from initial RN ed grouped

EDU_ENRLD

Enrolled upgrade ed program

GRAD_WHR

Graduated U.S. or non-U.S. RN program

NFINANCE

Finance means initial RN ed category count

NTRAINING

Emergency/disaster training areas category count

CH_CAREER

Employment change - career

CH_PERSFAM

Employment change - personal

CH_RETIRE

Employment change - retirement

CH_WRKPLCE

Employment change - workplace

CH_OTHER

Employment change - other

NCHREASN

Reasons position change category count

DOMFUNCT

Dominant function

EMP_STAT

Employment status 2008 detailed

EMPL_08

RN employment status 2008

EMPL_07

RN employment status 2007

EMP08_07

Employed in nursing 2008 & 2007

EMPL_TES

Employed by Temporary Employment Service (TES)

EMPL_ORG

Employed by organization/facility

EMPL_SELF

Self-employed in nursing

HRS_YR

Estimated hours per year for principal RN position

LST_WRKD

Years since paid RN

NOT_CAREER

Not working - career

NOT_PERSFAM

Not working - personal

NOT_RETIRE

Not working - retired

NOT_WORKPLACE

Not working - workplace

NOT_OTHER

Not working - other

NNOTRN

Reasons not working as RN category count

POSITION

Position title grouped

NLEVCARE

Levels of care principal RN job category count

NSPECIALTY

Clinical specialties principal RN job category count

SETSUM08

Employment setting summary 2008

SETSUM07

Employment setting summary 2007

SET07_08

RN employment setting 2007 versus 2008

TOTSAL

Total earnings from all nursing and non-nursing employment

NWKSTATE

Number of States in which RN works

NHLTHJOB

Health-related jobs pre-RN category count

APN_CNS

Clinical Nurse Specialist status

APN_NP

Nurse practitioner status

APN_NA

Nurse Anesthetist status

APN_NM

Nurse midwife status

APN_COMBOS

APRN status combinations

Q21_FIPSST

FIPS state 2008 principal RN position

Q21_DIV

Census division 2008 principal RN position

Q59_FIPSST

FIPS State 2008 residence

Q59_DIV

Census division 2008 residence

G_ST08

Geocode for State on 3/10/2008

G_RG08

Geocode for Census Division on 3/10/2008

RN_POP

RN working/living in U.S. 2008

RKRNWGTA

Raked weight

RKRNWGTA1

Raked weight 1

RKRNWGTA2

Raked weight 2

RKRNWGTA3

Raked weight 3

RKRNWGTA4

Raked weight 4

RKRNWGTA5

Raked weight 5

RKRNWGTA6

Raked weight 6

RKRNWGTA7

Raked weight 7

RKRNWGTA8

Raked weight 8

RKRNWGTA9

Raked weight 9

RKRNWGTA10

Raked weight 10

RKRNWGTA11

Raked weight 11

RKRNWGTA12

Raked weight 12

RKRNWGTA13

Raked weight 13

RKRNWGTA14

Raked weight 14

RKRNWGTA15

Raked weight 15

RKRNWGTA16

Raked weight 16

RKRNWGTA17

Raked weight 17

RKRNWGTA18

Raked weight 18

RKRNWGTA19

Raked weight 19

RKRNWGTA20

Raked weight 20

RKRNWGTA21

Raked weight 21

RKRNWGTA22

Raked weight 22

RKRNWGTA23

Raked weight 23

RKRNWGTA24

Raked weight 24

RKRNWGTA25

Raked weight 25

RKRNWGTA26

Raked weight 26

RKRNWGTA27

Raked weight 27

RKRNWGTA28

Raked weight 28

RKRNWGTA29

Raked weight 29

RKRNWGTA30

Raked weight 30

RKRNWGTA31

Raked weight 31

RKRNWGTA32

Raked weight 32

RKRNWGTA33

Raked weight 33

RKRNWGTA34

Raked weight 34

RKRNWGTA35

Raked weight 35

RKRNWGTA36

Raked weight 36

RKRNWGTA37

Raked weight 37

RKRNWGTA38

Raked weight 38

RKRNWGTA39

Raked weight 39

RKRNWGTA40

Raked weight 40

RKRNWGTA41

Raked weight 41

RKRNWGTA42

Raked weight 42

RKRNWGTA43

Raked weight 43

RKRNWGTA44

Raked weight 44

RKRNWGTA45

Raked weight 45

RKRNWGTA46

Raked weight 46

RKRNWGTA47

Raked weight 47

RKRNWGTA48

Raked weight 48

RKRNWGTA49

Raked weight 49

RKRNWGTA50

Raked weight 50

RKRNWGTA51

Raked weight 51

RKRNWGTA52

Raked weight 52

RKRNWGTA53

Raked weight 53

RKRNWGTA54

Raked weight 54

RKRNWGTA55

Raked weight 55

RKRNWGTA56

Raked weight 56

RKRNWGTA57

Raked weight 57

RKRNWGTA58

Raked weight 58

RKRNWGTA59

Raked weight 59

RKRNWGTA60

Raked weight 60

RKRNWGTA61

Raked weight 61

RKRNWGTA62

Raked weight 62

RKRNWGTA63

Raked weight 63

RKRNWGTA64

Raked weight 64

RKRNWGTA65

Raked weight 65

RKRNWGTA66

Raked weight 66

RKRNWGTA67

Raked weight 67

RKRNWGTA68

Raked weight 68

RKRNWGTA69

Raked weight 69

RKRNWGTA70

Raked weight 70

RKRNWGTA71

Raked weight 71

RKRNWGTA72

Raked weight 72

RKRNWGTA73

Raked weight 73

RKRNWGTA74

Raked weight 74

RKRNWGTA75

Raked weight 75

RKRNWGTA76

Raked weight 76

RKRNWGTA77

Raked weight 77

RKRNWGTA78

Raked weight 78

RKRNWGTA79

Raked weight 79

RKRNWGTA80

Raked weight 80

RKRNWGTA81

Raked weight 81

RKRNWGTA82

Raked weight 82

RKRNWGTA83

Raked weight 83

RKRNWGTA84

Raked weight 84

RKRNWGTA85

Raked weight 85

RKRNWGTA86

Raked weight 86

RKRNWGTA87

Raked weight 87

RKRNWGTA88

Raked weight 88

RKRNWGTA89

Raked weight 89

RKRNWGTA90

Raked weight 90

RKRNWGTA91

Raked weight 91

RKRNWGTA92

Raked weight 92

RKRNWGTA93

Raked weight 93

RKRNWGTA94

Raked weight 94

RKRNWGTA95

Raked weight 95

RKRNWGTA96

Raked weight 96

RKRNWGTA97

Raked weight 97

RKRNWGTA98

Raked weight 98

RKRNWGTA99

Raked weight 99

RKRNWGTA100

Raked weight 100

Q2_DIV

Census division State of first RN license

Q5_FIPSST

FIPS State initial RN ed program

Q5_DIV

Census division initial RN ed program

REG_B_ED

Census division of initial RN ed

REG_EMPL

Census division of 2008 employment

REG_RESI

Census division of 2008 residence

Q55_FIPSST

FIPS state 2007 principal RN position

Q55_DIV

Census division State of 2007 RN position

Q61_FIPSST

FIPS State 2007 residence

Q61_DIV

Census division 2007 residence

G_ST07

Geocode for State on 3/10/2007

G_RG07

Geocode for Census Division on 3/10/2007

G_MSA07

Geocode for metro/non-metro on 3/10/2007

G_MSA08

Geocode for metro/non-metro on 3/10/2008

GD_GEOST

State of graduations v. current GEO State

RCTY_SME

County of residence in 2007 and 2008

RST_SME

State of residence in 2007 and 2008

RCCC4Q21

Standard four category RUCA code - Q21

RCCC3Q21

Standard three category RUCA code - Q21

RCCC4Q55

Standard four category RUCA code - Q55

RCCC3Q55

Standard three category RUCA code - Q55

RCCC4Q59

Standard four category RUCA code - Q59

RCCC3Q59

Standard three category RUCA code - Q59

RCCC4Q61

Standard four category RUCA code - Q61

RCCC3Q61

Standard three category RUCA code - Q61

Q5COU_CONT

Continent initial RN ed program located

NLC

RN Licensure Compact

TOTRNHRS_YR

Total RN hours worked per year

CERT_ANST

Anesthesiology certification

CERT_ADDI

Addictions certification

CERT_ADMN

Administrator certification

CERT_AMBU

Ambulatory Care certification

CERT_AMBS

Ambulatory Surgical certification

CERT_CARD

Cardiac Rehabilitation certification

CERT_CASE

Case Management certification

CERT_COLL

College Health certification

CERT_COMM

Community Health certification

CERT_CNED

Continuing Ed/Staff Dev certification

CERT_COUN

Counselor certification

CERT_CRIT

Critical Care certification

CERT_DIAB

Diabetes Educator certification

CERT_FAMI

Family Practice certification

CERT_FLGT

Flight certification

CERT_GAST

Gastroenterology certification

CERT_GENN

General Practice certification

CERT_GENS

General Surgery certification

CERT_GERO

Gerontology certification

CERT_HEMO

Hemodialysis certification

CERT_HOLI

Holistic Health certification

CERT_HOME

Home Health certification

CERT_HOSP

Hospice certification

CERT_INFC

Infection Control certification

CERT_INFD

Infectious Disease certification

CERT_INFO

Informatics certification

CERT_INFU

Infusion Therapy certification

CERT_LACT

Lactation Consult certification

CERT_LEGL

Legal Nurse certification

CERT_MATN

Maternal Newborn certification

CERT_MEDS

Medical-Surgical certification

CERT_NEIN

Neonatal Intensive care certification

CERT_NELO

Neonatal, Low Risk certification

CERT_NEPH

Nephrology certification

CERT_NEUR

Neuroscience certification

CERT_NUED

Nurse Educator certification

CERT_OBST

Obstetric, Inpatient certification

CERT_OCCH

Occupational Health certification

CERT_ONCO

Oncology certification

CERT_ORTH

Orthopedic certification

CERT_PAIN

Pain Management certification

CERT_PEDI

Pediatrics certification

CERT_PERI

Perinatal certification

CERT_PSYC

Psych/Mental Health certification

CERT_QUAL

Quality Care certification

CERT_RADI

Radiological certification

CERT_FRST

RNs First Asst certification

CERT_REHA

Rehabilitation certification

CERT_RSCH

Research certification

CERT_SCHL

School certification

CERT_TRANS

Transplantation certification

CERT_UROL

Urology certification

CERT_WOMN

Women's Health certification

CERT_WOUN

Wound Care certification

CERT_TELE

Telemetry certification

OASIAN

Other Asian (non-Filipino) language

NSKLCERT

Skill-based certification category count

NAPNCERT

Advanced practice RN certification category count

NNTLCERT

National nursing certification category count

Q3_PUF

Initial RN degree/credential for PUF

PREP_SUM_PUF

Highest RN/RN-related ed summary for PUF

LI_RESST

State of first license versus GEO State for 2008

LICST_CT

Number of States in which RN currently licensed

ID_GENERAL

Randomly assigned ID for general file

IMPQ29C_1

Imputation flag for Q29C_1

IMPQ29C_2

Imputation flag for Q29C_2

IMPQ29C_3

Imputation flag for Q29C_3

IMPQ29C_4

Imputation flag for Q29C_4

IMPQ29C_5

Imputation flag for Q29C_5

IMPQ29C_6

Imputation flag for Q29C_6

IMPQ29C_7

Imputation flag for Q29C_7

IMPQ29C_8

Imputation flag for Q29C_8

IMPQ29C_9

Imputation flag for Q29C_9

IMPQ29C_10

Imputation flag for Q29C_10

IMPQ29C_11

Imputation flag for Q29C_11

IMPQ29C_12

Imputation flag for Q29C_12

IMPQ29C_13

Imputation flag for Q29C_13

IMPQ29C_14

Imputation flag for Q29C_14

IMPQ29C_15

Imputation flag for Q29C_15

IMPQ29C_16

Imputation flag for Q29C_16

IMPQ29C_17

Imputation flag for Q29C_17

IMPQ29C_18

Imputation flag for Q29C_18

IMPQ29C_19

Imputation flag for Q29C_19

IMPQ29C_20

Imputation flag for Q29C_20

IMPQ29C_21

Imputation flag for Q29C_21

IMPQ29C_22

Imputation flag for Q29C_22

IMPQ29C_23

Imputation flag for Q29C_23

IMPQ29C_24

Imputation flag for Q29C_24

IMPQ29C_25

Imputation flag for Q29C_25

IMPQ29C_26

Imputation flag for Q29C_26

IMPQ29C_27

Imputation flag for Q29C_27

IMPQ29C_28

Imputation flag for Q29C_28

IMPQ29C_29

Imputation flag for Q29C_29

IMPQ29C_30

Imputation flag for Q29C_30

IMPQ29A_1

Imputation flag for Q29A_1

IMPQ29A_2

Imputation flag for Q29A_2

IMPQ29A_3

Imputation flag for Q29A_3

IMPQ29A_4

Imputation flag for Q29A_4

IMPQ29A_5

Imputation flag for Q29A_5

IMPQ29A_6

Imputation flag for Q29A_6

IMPQ29A_7

Imputation flag for Q29A_7

IMPQ29A_8

Imputation flag for Q29A_8

IMPQ29A_9

Imputation flag for Q29A_9

IMPQ29A_10

Imputation flag for Q29A_10

IMPQ29A_11

Imputation flag for Q29A_11

IMPQ29A_12

Imputation flag for Q29A_12

IMPQ29A_13

Imputation flag for Q29A_13

IMPQ29A_14

Imputation flag for Q29A_14

IMPQ29A_15

Imputation flag for Q29A_15

IMPQ29A_16

Imputation flag for Q29A_16

IMPQ29A_17

Imputation flag for Q29A_17

IMPQ29A_18

Imputation flag for Q29A_18

IMPQ29A_19

Imputation flag for Q29A_19